+++ commands/build.js
@@ -6,8 +6,8 @@ const chalk = require('chalk');
 const createWebpackConfig = require('../lib/create-webpack-config');
 const logger = require('../lib/logger');
 const publicFilesCopier = require('../lib/public-files-copier');
-const htmlCompiler = require('../lib/html-compiler');
-const cssCompiler = require('../lib/css-compiler');
+const { writeHtml } = require('../lib/html-compiler');
+const { writeCss } = require('../lib/css-compiler');
 const writeWebpackStats = require('../lib/write-webpack-stats');
 const webpackPromise = require('../lib/webpack-promise');
 
@@ -24,9 +24,9 @@ function build(urc) {
         writeWebpackStats(urc.stats, stats);
       }
     })
-    .then(() => cssCompiler.write(urc))
-    .then(cssFilename => htmlCompiler.write(urc, cssFilename))
-    .then(() => publicFilesCopier.copy(urc))
+    .then(() => writeCss(urc))
+    .then(cssFilename => writeHtml(urc, cssFilename))
+    .then(() => publicFilesCopier(urc))
     .then(() => {
       // Clean up files you won't need to deploy.
       del.sync(path.join(urc.outputDirectory, 'assets.json'));
