+++ lib/css-compiler.js
@@ -6,6 +6,10 @@ const autoprefixer = require('autoprefixer');
 const postcssConcatenator = require('./packageable/postcss-concatenator');
 const { CSS_BASENAME } = require('./constants');
 
+module.exports = {
+  writeCss
+};
+
 function writeCss(urc) {
   if (urc.stylesheets.length === 0) {
     return Promise.resolve();
@@ -21,7 +25,3 @@ function writeCss(urc) {
     hash: urc.production
   });
 }
-
-module.exports = {
-  writeCss
-};
