+++ lib/html-compiler.js
@@ -8,6 +8,8 @@ const pify = require('pify');
 
 const { CSS_BASENAME } = require('./constants');
 
+module.exports = { renderHtml, writeHtml };
+
 function renderHtml(urc, cssFilename) {
   const assets = urc.getAssets();
   // Inline the runtime JS because it's super small.
@@ -71,10 +73,7 @@ function uglifyRuntime(urc, assets) {
 }
 
 function writeHtml(urc, cssFilename) {
-  // TOFIX renderHtml takes cssFilename param?
   return renderHtml(urc, cssFilename).then(html =>
     pify(fs.writeFile)(path.join(urc.outputDirectory, 'index.html'), html)
   );
 }
-
-module.exports = { renderHtml, writeHtml };
