+++ lib/public-files-copier.js
@@ -1,7 +1,8 @@
 'use strict';
 
 const autoCopy = require('./packageable/auto-copy');
-const logger = require('./logger');
+
+module.exports = copyPublicFiles;
 
 function copyPublicFiles(urc) {
   return autoCopy.copy({
@@ -9,22 +10,3 @@ function copyPublicFiles(urc) {
     destDir: urc.outputDirectory
   });
 }
-
-function watchPublicFiles(urc) {
-  const copier = autoCopy.watch({
-    sourceDir: urc.publicDirectory,
-    destDir: urc.outputDirectory
-  });
-  copier.on('copy', filename => {
-    logger.log(`Copying ${filename}`);
-  });
-  copier.on('delete', filename => {
-    logger.log(`Deleting ${filename}`);
-  });
-  copier.on('error', logger.error);
-}
-
-module.exports = {
-  copy: copyPublicFiles,
-  watch: watchPublicFiles
-};
